Fix misplaced dashboard tests and removed duplicate tests
