+++ tests/test_account.py
@@ -1,4 +1,3 @@
-import json
 import os
 import uuid
 from unittest.mock import patch
@@ -212,19 +211,6 @@ def test_user_ajax_label(customer_user):
 
 def test_user_ajax_label_without_address(admin_user):
     assert admin_user.get_ajax_label() == admin_user.email
-
-
-def test_ajax_users_list(admin_client, admin_user, customer_user):
-    users_list = [
-        {'id': admin_user.pk, 'text': admin_user.get_ajax_label()},
-        {'id': customer_user.pk, 'text': customer_user.get_ajax_label()}]
-    url = reverse('dashboard:ajax-users-list')
-
-    response = admin_client.get(url, HTTP_X_REQUESTED_WITH='XMLHttpRequest')
-    resp_decoded = json.loads(response.content.decode('utf-8'))
-
-    assert response.status_code == 200
-    assert resp_decoded == {'results': users_list}
 
 
 def test_disabled_recaptcha():
